requirements.txt
torch>=2.0
torchvision>=0.15
transformers>=4.37
numpy>=1.24
pandas>=2.0
scikit-learn>=1.2
tqdm>=4.65
Pillow>=9.5
